requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-mysql==2.1.1
djangorestframework==3.7.1
idna==2.6
kombu==4.1.0
mysqlclient==1.3.12
pytz==2017.2
requests==2.18.4
urllib3==1.22
vine==1.1.4
